Ejected, added syntax coloring in nw app.
I'm really sad to admit it, but I had to eject for this. I spent month
trying to find a way to do this (if something like create-nw-react-app-rewired
existed i could have done it without ejecting), but sadly it seems
like it's not possible.

This won't impact the web version, which is using create-react-app-rewired.